requirements.txt
asttokens==2.4.1
certifi==2024.2.2
charset-normalizer==3.3.2
decorator==5.1.1
exceptiongroup==1.2.1
executing==2.0.1
filelock==3.13.4
fsspec==2024.3.1
idna==3.7
ipython==8.23.0
jedi==0.19.1
Jinja2==3.1.3
llvmlite==0.42.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
more-itertools==10.2.0
mpmath==1.3.0
networkx==3.3
numba==0.59.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab
parso==0.8.4
pexpect==4.9.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
regex==2024.4.16
requests==2.31.0
six==1.16.0
stack-data==0.6.3
sympy==1.12
tiktoken==0.6.0
torch==2.2.2
tqdm==4.66.2
traitlets==5.14.3
triton==2.2.0
typing_extensions==4.11.0
urllib3==2.2.1
wcwidth==0.2.13
